Clear reindex flag when indexing an existing item
The search indexing could get stuck on processing the same content
repeatedly because items marked for reindexing were not being unmarked
afterwards.

Change the database query that is run for indexed items to also cover
this case and properly clear the reindex flag as part of indexing an
existing item.

Note: the used query is MySQL/MariaDB specific and not standard SQL.